pyproject.toml
[project]
name = "fixibake"
version = "0.1.0"
requires-python = "~=3.11.0"
dependencies = [
    "python-magic-bin",
    "pyinstaller",
    "tkinterdnd2",
    "wordfreq[cjk]",
]
